requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

asyncpg==0.30.0
redis==5.2.1

google-generativeai==0.8.3

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

python-dotenv==1.0.1
geopy==2.4.1
geohash2==1.1
